requirements.txt
aiogram==3.11.2
python-dotenv==1.0.1
aiofiles==23.2.0
openpyxl==3.1.2
pandas==2.0.3
google-api-python-client==2.108.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0